Refactor console
